Move App.js from Components folder
